platform-api/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pyyaml==6.0.2
requests==2.32.3
python-multipart==0.0.20
pytest==8.3.5
httpx==0.28.1
